ref(compiler): split ConstantFolder and FnSymbol assert rewriting
Mirrors the CallSpecialization precedent. Pure refactor; fold return
routing and pre/post assert expansion are behavior-identical.

- ConstantFolder becomes a dispatcher delegating to LiteralArithmeticFolder
  (scalar + - * inc dec = < ... with overflow guards), LiteralCollectionFolder
  (count/first/last/nth/reduce/str literal folding, driving reduce through the
  arithmetic folder), and LiteralBitwiseFolder. foldIf and the single-literal
  bool predicates stay on the dispatcher.
- FnSymbol delegates {:pre :post} -> assert rewriting (including the recursive
  %-replacement walk and the leading-let unwrap) to FnPrePostConditionRewriter,
  which now owns the $assertsEnabled flag. FnSymbol's public constructor is
  unchanged.

Related to #2264
