+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 - Compiler internals: split the 1325-LOC `CallSpecialization` god-class into eight focused collaborators (pure refactor, output unchanged)
 - Architecture: split three multi-responsibility classes into focused stateless collaborators (pure refactor, behavior unchanged) (#2262): `DefSymbol` delegates macro `&form`/`&env` + `^:tag` form rewriting to `MacroFormRewriter` and `:arglists` formatting to `DefArglistBuilder`; `ParamTypeInferrer` delegates PHP-/`phel.core`-call shape analysis to `CallTypeExpectationResolver`; `TestCommand` delegates option parsing/coercion + the parallelism decision to `TestCommandOptionParser`
 - Architecture: promoted two implicit cross-module contracts to their proper homes (pure relocations, behavior unchanged) (#2263): the parse-tree value objects (`Compiler\Domain\Parser\ParserNode\*` and the lexer `Token`) moved to `Phel\Shared\Parser\Node\*` (consumed as an AST contract by Formatter/Lint/Api while Compiler still produces them; removes the `CompilerFacadeInterface -> Compiler\Domain` parse-tree references), and the stateful `LoadClasspath` moved from `Compiler\Domain\Analyzer\Resolver` to `Phel\Lang`, next to the `Registry` slot it owns
+- Compiler internals: split the 882-LOC `ConstantFolder` into a thin dispatcher plus `LiteralArithmeticFolder`, `LiteralCollectionFolder`, and `LiteralBitwiseFolder`, and extracted `FnSymbol`'s `{:pre :post}` assert rewriting into `FnPrePostConditionRewriter` (pure refactor, fold routing and assert expansion unchanged) (#2264)
 - Compiler internals: centralized the emitters' bare-expression capture into `OutputEmitter::captureNodeAsExpression()`
 - Tests: `RegistryTest` and `PhelVarTest` snapshot and restore the global `Registry`, fixing order-dependent `phel.core` suite failures (#2256)
 - Docs: condensed every `docs/` guide (~17% smaller), verified against the runtime, and cross-linked to phel-lang.org
